feat: allow cancelling job runs in the pipeline editor
