+++ services/orchest-webserver/client/src/pipeline-view/pipeline-canvas-header-bar/PipelineCanvasHeaderBar.tsx
@@ -43,16 +43,13 @@ export const PipelineCanvasHeaderBar = () => {
           Logs
         </Button>
         <ServicesMenu />
-        {!isJobRun && (
-          <>
-            <Divider
-              orientation="vertical"
-              sx={{ height: (theme) => theme.spacing(3) }}
-            />
-            <CreateStepButton />
-            <PrimaryPipelineButton />
-          </>
-        )}
+
+        <Divider
+          orientation="vertical"
+          sx={{ height: (theme) => theme.spacing(3) }}
+        />
+        {!isJobRun && <CreateStepButton />}
+        <PrimaryPipelineButton />
         <PipelineMoreOptionsMenu />
       </Stack>
     </Stack>
